Modify test, extract password rule to its own file and clean up
