+++ packages/dd-trace/src/appsec/iast/analyzers/hardcoded-password-analyzer.js
@@ -3,15 +3,7 @@
 const { HARDCODED_PASSWORD } = require('../vulnerabilities')
 const HardcodedBaseAnalyzer = require('./hardcoded-base-analyzer')
 
-const allRules = [
-  {
-    id: 'hardcoded-password',
-
-    // eslint-disable-next-line max-len
-    regex: /(?:pwd|pswd|pass|secret)(?:[0-9a-z\-_\t.]{0,20})(?:[\s|']|[\s|""]){0,3}(?:=|>|:{1,3}=|\|\|:|<=|=>|:|\?=)(?:'|""|\s|=|\x60){0,5}([0-9a-z\-_.=]{10,150})(?:['"\s\x60;]|$)/i,
-    type: 'NameAndValue'
-  }
-]
+const allRules = require('./hardcoded-password-rules')
 
 class HardcodedPasswordAnalyzer extends HardcodedBaseAnalyzer {
   constructor () {
