+++ packages/dd-trace/test/appsec/iast/analyzers/hardcoded-password-analyzer.spec.js
@@ -11,21 +11,18 @@ const Config = require('../../../../src/config')
 const hardcodedPasswordAnalyzer = require('../../../../src/appsec/iast/analyzers/hardcoded-password-analyzer')
 const iast = require('../../../../src/appsec/iast')
 
-const suite = [
-  {
-    id: 'hardcoded-password',
-    samples: [
-      { ident: 'hashpwd', value: 'hpu0-ig=3o5slyr0rkqszidgxw-bc23tivq8e1-qvt.4191vlwm8ddk.ce64m4q0kga' },
-      { ident: 'passphrase', value: '8jqqn=-5k9fnu4ukjzg5ar-=yw4r3_no=cn25zpegfh3.ndid5lp04iknqwvz92npuniksena2sz9w1bipd_g_oyx2ij3xi7cyh8=y.pv_f_gnbupl' },
-      { ident: 'password', value: '9z57r5sph9zkgnknefwbnrr5ilqiavsylk9b6qt2a9kg-g=ez=_fwat' },
-      { ident: 'passw', value: '8=t6jaeqk0=-e9gspu0e6b91mcmayd2bvq37_nn0yyzzh20qn8t29eg-trn6r4w7.s4r59u160jcgenjb-rpn=1ga' },
-      { ident: 'secret', value: 'ha=gcu1r-t.ckeirz9y3jf34zvg7o.qsh.' },
-      { ident: 'passkey', value: 'k2urtbthyl=fd6z4wl6r26zlldy3.39ymm.p4l5wu_9mzg30sxx6absd696fgpjub4wu8a0bge-5_dp59xf493oayojzftf4iiavndwixmt-fngxn05naek8' },
-      { ident: 'pass', value: 'm-ou4zr=vri2yl-0_ardsza7qbenvy3_2b1h8rsq2_n-.utj9nd3xyvqpg4xl37-nl0nkjwam1avbe9zl' },
-      { ident: 'pwd', value: 'e8l=47jevvmz5=trchu6.uu3lwn0fb79bpt=fogw36r1srzb1o1w4f-nhihcni=kncnj9ubs0.2w2tey-b=u9f4-s5l_648p67hf4f2bccv6c4lr3mfcj8a77qv38dlzuhpyb=u' },
-      { ident: 'pswd', value: '9rh_n.oooxynt-6._a7ho.2=v-gziqf3wz2vudn916jej_6xaq_1vczboi5rmt5_iuvxxf8oq.ghisjxum' },
-      { ident: 'passwd', value: 'vtecj=v6b7-qc1m-s6c=zidew-hw-=c-=4d83icy28-guc3g-vvrimsdf=jml.acy=q7sdwaxh_rl-okx1z48pihg=w4=tc4' }]
-  }
+const ruleId = 'hardcoded-password'
+const samples = [
+  { ident: 'hashpwd', value: 'hpu0-ig=3o5slyr0rkqszidgxw-bc23tivq8e1-qvt.4191vlwm8ddk.ce64m4q0kga' },
+  { ident: 'passphrase', value: '8jqqn=-5k9fnu4ukjzg5ar-=yw4r3_no=cn25zpegfh3.ndid5lp04iknqwvz92npuniksena2sz9w1bipd_g_oyx2ij3xi7cyh8=y.pv_f_gnbupl' },
+  { ident: 'password', value: '9z57r5sph9zkgnknefwbnrr5ilqiavsylk9b6qt2a9kg-g=ez=_fwat' },
+  { ident: 'passw', value: '8=t6jaeqk0=-e9gspu0e6b91mcmayd2bvq37_nn0yyzzh20qn8t29eg-trn6r4w7.s4r59u160jcgenjb-rpn=1ga' },
+  { ident: 'secret', value: 'ha=gcu1r-t.ckeirz9y3jf34zvg7o.qsh.' },
+  { ident: 'passkey', value: 'k2urtbthyl=fd6z4wl6r26zlldy3.39ymm.p4l5wu_9mzg30sxx6absd696fgpjub4wu8a0bge-5_dp59xf493oayojzftf4iiavndwixmt-fngxn05naek8' },
+  { ident: 'pass', value: 'm-ou4zr=vri2yl-0_ardsza7qbenvy3_2b1h8rsq2_n-.utj9nd3xyvqpg4xl37-nl0nkjwam1avbe9zl' },
+  { ident: 'pwd', value: 'e8l=47jevvmz5=trchu6.uu3lwn0fb79bpt=fogw36r1srzb1o1w4f-nhihcni=kncnj9ubs0.2w2tey-b=u9f4-s5l_648p67hf4f2bccv6c4lr3mfcj8a77qv38dlzuhpyb=u' },
+  { ident: 'pswd', value: '9rh_n.oooxynt-6._a7ho.2=v-gziqf3wz2vudn916jej_6xaq_1vczboi5rmt5_iuvxxf8oq.ghisjxum' },
+  { ident: 'passwd', value: 'vtecj=v6b7-qc1m-s6c=zidew-hw-=c-=4d83icy28-guc3g-vvrimsdf=jml.acy=q7sdwaxh_rl-okx1z48pihg=w4=tc4' }
 ]
 
 describe('Hardcoded Password Analyzer', () => {
@@ -42,24 +39,22 @@ describe('Hardcoded Password Analyzer', () => {
 
     afterEach(sinon.restore)
 
-    suite.forEach((testCase) => {
-      testCase.samples.forEach((sample, sampleIndex) => {
-        // sample values are arrays containing the parts of the original token
-        it(`should match rule ${testCase.id} with #${sampleIndex + 1} value ${sample.ident}...`, () => {
-          hardcodedPasswordAnalyzer.analyze({
-            file,
-            literals: [{
-              value: sample.value,
-              locations: [{
-                ident: sample.ident,
-                line,
-                column
-              }]
+    samples.forEach((sample, sampleIndex) => {
+      // sample values are arrays containing the parts of the original token
+      it(`should match rule ${ruleId} with #${sampleIndex + 1} value ${sample.ident}...`, () => {
+        hardcodedPasswordAnalyzer.analyze({
+          file,
+          literals: [{
+            value: sample.value,
+            locations: [{
+              ident: sample.ident,
+              line,
+              column
             }]
-          })
-
-          expect(report).to.have.been.calledOnceWithExactly({ file: relFile, line, column, data: testCase.id })
+          }]
         })
+
+        expect(report).to.have.been.calledOnceWithExactly({ file: relFile, line, column, data: ruleId })
       })
     })
